+++ __tests__/integration/rawCaptureCreated.spec.ts
@@ -1,9 +1,9 @@
 import log from 'loglevel'
 import { createBroker, publish } from 'messaging/broker'
 import brokerConfig from 'messaging/brokerConfig'
+import registerEventHandlers from 'messaging/eventHandlers'
 import { truncateTables } from 'models/base'
 import { BrokerAsPromised, withTestConfig } from 'rascal'
-import registerEventHandlers from 'services/eventHandlers'
 import knex, { TableNames } from 'services/knex'
 
 import capture_in_kenya from '../mock/capture_in_kenya.json'
